requirements.txt
matplotlib==3.8.2
numpy==1.24.1
pandas==2.1.4
Pillow==10.1.0
scikit_learn==1.3.2
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.36.2
